requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
PyYAML==6.0.1
requests==2.31.0
reportlab==4.0.4
python-magic==0.4.27
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
python-dateutil==2.8.2
urllib3==1.26.18